Validation, Refactoring, String Accout Numbers, Transfer Reference Numbers

+++ backend/RetailBank/Services/ILoanService.cs
@@ -2,7 +2,7 @@ namespace RetailBank.Services;
 
 public interface ILoanService
 {
-    public Task<ulong> CreateLoanAccount(ulong debitAccountNumber, ulong loanAmount);
-    public Task ChargeInterest(ulong loanAccountId);
-    public Task PayInstallment(ulong loanAccountId);
+    public Task<UInt128> CreateLoanAccount(UInt128 debitAccountNumber, ulong loanAmount);
+    public Task ChargeInterest(UInt128 loanAccountId);
+    public Task PayInstallment(UInt128 loanAccountId);
 }
